requirements.txt
PyYAML
moviepy
Pillow